暖炉に近づいたらパチパチなるように対応

+++ src/CameraTest.hpp
@@ -1277,5 +1277,5 @@ private:
 	float debugHeight = 200;
 
 	int footpattern = 0;
-	int footcount = 0;
+	float footcount = 0;
 };
